# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/demandeInscript/store": {
            "post": {
                "tags": [
                    "Demande Inscriptions"
                ],
                "summary": "Create a new Demande Inscription",
                "description": "Create a new Demande Inscription",
                "operationId": "storeDemandeInscription",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DemandeInscription"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DemandeInscription"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DemandeInscription": {
                "title": "Demande Inscription",
                "description": "Demande Inscription model",
                "properties": {
                    "nomEcole": {
                        "type": "string"
                    },
                    "adresseEcole": {
                        "type": "string"
                    },
                    "descriptionEcole": {
                        "type": "string"
                    },
                    "imageEcole": {
                        "type": "string"
                    },
                    "nomA": {
                        "type": "string"
                    },
                    "prenomA": {
                        "type": "string"
                    },
                    "emailA": {
                        "type": "string"
                    },
                    "cin": {
                        "type": "string"
                    },
                    "numTel": {
                        "type": "string"
                    },
                    "imageA": {
                        "type": "string"
                    },
                    "dateNaissance": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}